Fixed StringTemplate4 to work with ARC and pass all but one test in TestLineWrap.

+++ objc/main/tests/objc/TestGroups.h
@@ -40,7 +40,7 @@
 @end
 
 @interface Counter : NSObject {
-  int n;
+  NSInteger n;
 }
 
 - (id) init;
